# valhalla/jawn/src/tsoa-build/private/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Result_null.string_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_null_"
					},
					{
						"$ref": "#/components/schemas/ResultError_string_"
					}
				]
			},
...
			"Result_string.string_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_string_"
					},
					{
						"$ref": "#/components/schemas/ResultError_string_"
					}
				]
			},
...
			"Result_Result__color-string--created_at-string--domain-string--governance_settings-Json--has_onboarded-boolean--icon-string--id-string--is_main_org-boolean--is_personal-boolean--limits-Json--logo_path-string--name-string--onboarding_status-Json--org_provider_key-string--organization_type-string--owner-string--percent_to_log-number--referral-string--request_limit-number--reseller_id-string--size-string--soft_delete-boolean--stripe_customer_id-string--stripe_metadata-Json--stripe_subscription_id-string--stripe_subscription_item_id-string--subscription_status-string--tier-string_-Array.string_.unknown_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_Result__color-string--created_at-string--domain-string--governance_settings-Json--has_onboarded-boolean--icon-string--id-string--is_main_org-boolean--is_personal-boolean--limits-Json--logo_path-string--name-string--onboarding_status-Json--org_provider_key-string--organization_type-string--owner-string--percent_to_log-number--referral-string--request_limit-number--reseller_id-string--size-string--soft_delete-boolean--stripe_customer_id-string--stripe_metadata-Json--stripe_subscription_id-string--stripe_subscription_item_id-string--subscription_status-string--tier-string_-Array.string__"
					},
					{
						"$ref": "#/components/schemas/ResultError_unknown_"
					}
				]
			},
			"NewOrganizationParams": {
				"properties": {
					"tier": {
						"type": "string",
						"nullable": true
					},
					"subscription_status": {
						"type": "string",
						"nullable": true
					},
					"stripe_subscription_item_id": {
						"type": "string",
						"nullable": true
					},
					"stripe_subscription_id": {
						"type": "string",
						"nullable": true
					},
					"stripe_metadata": {
						"$ref": "#/components/schemas/Json"
					},
					"stripe_customer_id": {
						"type": "string",
						"nullable": true
					},
					"soft_delete": {
						"type": "boolean"
					},
					"size": {
						"type": "string",
						"nullable": true
					},
					"reseller_id": {
						"type": "string",
						"nullable": true
					},
					"request_limit": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"referral": {
						"type": "string",
						"nullable": true
					},
					"percent_to_log": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"owner": {
						"type": "string"
					},
					"organization_type": {
						"type": "string"
					},
					"org_provider_key": {
						"type": "string",
						"nullable": true
					},
					"onboarding_status": {
						"$ref": "#/components/schemas/Json"
					},
					"name": {
						"type": "string"
					},
					"logo_path": {
						"type": "string",
						"nullable": true
					},
					"limits": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Json"
							}
						],
						"nullable": true
					},
					"is_personal": {
						"type": "boolean"
					},
					"is_main_org": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"has_onboarded": {
						"type": "boolean"
					},
					"governance_settings": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Json"
							}
						],
						"nullable": true
					},
					"domain": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"nullable": true
					},
					"color": {
						"type": "string"
					}
				},
				"required": [
					"owner",
					"name"
				],
				"type": "object"
			},
...
			"UpdateOrganizationParams": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_NewOrganizationParams.name-or-color-or-icon-or-org_provider_key-or-limits-or-reseller_id-or-organization_type-or-onboarding_status_"
					},
					{
						"properties": {
							"variant": {
								"type": "string"
							}
						},
						"type": "object"
					}
				]
			},
			"UIFilterRowTree": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/UIFilterRowNode"
					},
					{
						"$ref": "#/components/schemas/FilterRow"
					}
				]
			},
...
			"OrganizationFilter": {
				"properties": {
					"softDelete": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"filter": {
						"items": {
							"$ref": "#/components/schemas/UIFilterRowTree"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"softDelete",
					"filter",
					"name",
					"id"
				],
				"type": "object"
			},
...
			"Result_OrganizationLayout.string_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_OrganizationLayout_"
					},
					{
						"$ref": "#/components/schemas/ResultError_string_"
					}
				]
			},
...
			"Result_OrganizationMember-Array.string_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_OrganizationMember-Array_"
					},
					{
						"$ref": "#/components/schemas/ResultError_string_"
					}
				]
			},
...
			"Result_OrganizationOwner-Array.string_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_OrganizationOwner-Array_"
					},
					{
						"$ref": "#/components/schemas/ResultError_string_"
					}
				]
			},
...
			"OnboardingStatus": {
				"$ref": "#/components/schemas/Partial__currentStep-string--selectedTier-string--hasOnboarded-boolean--members-any-Array--addons_58__prompts-boolean--experiments-boolean--evals-boolean___"
			},
...
			"Result_StoreFilterType-Array.string_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResultSuccess_StoreFilterType-Array_"
					},
					{
						"$ref": "#/components/schemas/ResultError_string_"
					}
				]
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "helicone-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/v1/organization/reseller/{resellerId}": {
			"get": {
				"operationId": "GetReseller",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_Result__color-string--created_at-string--domain-string--governance_settings-Json--has_onboarded-boolean--icon-string--id-string--is_main_org-boolean--is_personal-boolean--limits-Json--logo_path-string--name-string--onboarding_status-Json--org_provider_key-string--organization_type-string--owner-string--percent_to_log-number--referral-string--request_limit-number--reseller_id-string--size-string--soft_delete-boolean--stripe_customer_id-string--stripe_metadata-Json--stripe_subscription_id-string--stripe_subscription_item_id-string--subscription_status-string--tier-string_-Array.string_.unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "resellerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/organization/user/accept_terms": {
			"post": {
				"operationId": "AcceptTerms",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/v1/organization/create": {
			"post": {
				"operationId": "CreateNewOrganization",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_string.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewOrganizationParams"
							}
						}
					}
				}
			}
		},
		"/v1/organization/{organizationId}/update": {
			"post": {
				"operationId": "UpdateOrganization",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOrganizationParams"
							}
						}
					}
				}
			}
		},
		"/v1/organization/onboard": {
			"post": {
				"operationId": "OnboardOrganization",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/organization/{organizationId}/add_member": {
			"post": {
				"operationId": "AddMemberToOrganization",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/organization/{organizationId}/create_filter": {
			"post": {
				"operationId": "CreateOrganizationFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"filterType": {
										"type": "string",
										"enum": [
											"dashboard",
											"requests"
										]
									},
									"filters": {
										"items": {
											"$ref": "#/components/schemas/OrganizationFilter"
										},
										"type": "array"
									}
								},
								"required": [
									"filterType",
									"filters"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/organization/{organizationId}/update_filter": {
			"post": {
				"operationId": "UpdateOrganizationFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"filterType": {
										"type": "string",
										"enum": [
											"dashboard",
											"requests"
										]
									},
									"filters": {
										"items": {
											"$ref": "#/components/schemas/OrganizationFilter"
										},
										"type": "array"
									}
								},
								"required": [
									"filterType",
									"filters"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/organization/delete": {
			"delete": {
				"operationId": "DeleteOrganization",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/v1/organization/{organizationId}/layout": {
			"get": {
				"operationId": "GetOrganizationLayout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_OrganizationLayout.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filterType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/organization/{organizationId}/members": {
			"get": {
				"operationId": "GetOrganizationMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_OrganizationMember-Array.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/organization/{organizationId}/update_member": {
			"post": {
				"operationId": "UpdateOrganizationMember",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memberId": {
										"type": "string"
									},
									"role": {
										"type": "string"
									}
								},
								"required": [
									"memberId",
									"role"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/organization/{organizationId}/owner": {
			"get": {
				"operationId": "GetOrganizationOwner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_OrganizationOwner-Array.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/organization/{organizationId}/remove_member": {
			"delete": {
				"operationId": "RemoveMemberFromOrganization",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "memberId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/organization/setup-demo": {
			"post": {
				"operationId": "SetupDemo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/v1/organization/update_onboarding": {
			"post": {
				"operationId": "UpdateOnboardingStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Organization"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"has_onboarded": {
										"type": "boolean"
									},
									"name": {
										"type": "string"
									},
									"onboarding_status": {
										"$ref": "#/components/schemas/OnboardingStatus"
									}
								},
								"required": [
									"has_onboarded",
									"name",
									"onboarding_status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/v1/filter": {
			"get": {
				"operationId": "GetFilters",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_StoreFilterType-Array.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Filter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result__id-string_.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Filter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreFilterType"
							}
						}
					}
				}
			}
		},
		"/v1/filter/{id}": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_StoreFilterType.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Filter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Filter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_null.string_"
								}
							}
						}
					}
				},
				"tags": [
					"Filter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"filters": {}
								},
								"required": [
									"filters"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "https://api.helicone.ai/"
		},
		{
			"url": "http://localhost:8585/"
		}
	]
}